+++ packages/vendor-interoperability/src/mock-client/Base58Btc.ts
