+++ packages/vendor-interoperability/src/mock-client/MockInvoker.ts
@@ -1,5 +1,5 @@
 import forge from 'node-forge';
-import { decode } from './BadEncoding';
+import { decode } from './Base58Btc';
 import * as config from './config';
 
 const {
